Added example, and removed duplicates
Added example on how to get resource groups based on tags. Also, removed what was basically duplicates. There were two examples on how to search for resource groups with wildcards in their names. Removed the latest example, and rewrote the other to not use Where-Object as there are no reason to do so.